Fix broken tests with new is_active logic (see commit message for deets)
 - There were some test cases that were treating the activation
timestamps and deactivation timestamps as timestamps between the current
period.. but this is incorrect

For example, one test case had no activation or deactivation times but
the expected result was 31 days (full month), this is incorrect.

The other test that was updated had no activation time, but only a
deactivation time. This is not possible without something being very
wrong.